# Speeding up adding factors + computing wiring