… .pre-commit-config.yaml
… examples/gmrf.py
… examples/ising_model.py
… examples/pmp_binary_deconvolution.py
… examples/rbm.py
… examples/rcn.py
… pgmax/fg/groups.py
… pgmax/fg/nodes.py
… pgmax/groups/enumeration.py
… pgmax/groups/logical.py
… pgmax/groups/variables.py
… tests/factors/test_and.py
… tests/factors/test_or.py
… tests/fg/test_graph.py
… tests/fg/test_groups.py
… tests/fg/test_wiring.py
… tests/test_pgmax.py