Main goal of this PR is to speed up the process of adding factors to a FactorGraph and to compile wiring. In particular
- we do not instantiate factors in a FactorGraph by default. Instead all the computations, including wiring, happen at the factor groups level. 
- all the factors-related operations are computed on demand, when the user requires them
- we create a subfolder `pgmax/groups/` to clarify the structure

Before:
- adding factors for the RBM exp takes 25s, building run_bp takes 45s
- adding factors for the convor exp takes 20s, building run_bp takes 25s

After:
- adding factors for the RBM exp takes 2s, building run_bp takes 4s
- adding factors for the convor exp takes 2s, building run_bp takes 10s

Our next step will accelerate the wiring compilation further by  using `numba`