@antoine-dedieu 2022-04-01
Thanks for your thorough review @StannisZhou 
There are a few points we should discuss related to mypy behaviour
I totally agree that the next PR should be focused on speeding the factor group wiring using numba!